backend/pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["omee <wasif.omee@mbm.group>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework-simplejwt = "^5.5.0"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
channels = "^4.2.2"
channels-redis = "^4.2.1"
daphne = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
